+++ packages/elements/src/combo-box/__test__/utils.js
@@ -6,13 +6,14 @@ export const snapshotIgnore = {
 
 /**
  * Cross browser function to wait while select element becomes opened/closed and resized
+ * @async
  * @param {Select} el Select
- * @returns {void}
+ * @returns {Promise<void>} Empty promise
  */
 export const openedUpdated = async (el) => {
   await elementUpdated(el);
   await nextFrame();
-  await nextFrame(); // IE11 needs a second iframe, otherwise resize observer is not run;
+  await nextFrame(); // IE11 needs another frame, otherwise resize observer is not run;
 };
 
 export const data = [{
@@ -54,4 +55,11 @@ export const makeQueryRequest = async (el, textInput) => {
   await oneEvent(el, 'query-changed');
   await elementUpdated(el);
   await aTimeout(100);
+};
+
+export const dispatchCustomEvent = (el, eventName) => {
+  el.dispatchEvent(new CustomEvent(eventName, {
+    bubbles: true,
+    composed: true
+  }));
 };