+++ packages/elements/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.6.2](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.6.1...@refinitiv-ui/elements@6.6.2) (2023-02-14)
+
+### Bug Fixes
+
+* **combo-box:** fix value-changed event of free text mode ([#583](https://github.com/Refinitiv/refinitiv-ui/pull/583))  ([bae20f0](https://github.com/Refinitiv/refinitiv-ui/commit/bae20f0854d9f95d8c10985f8b2c5a9619922004))
+
+
+
+
+
+## [6.6.1](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.6.0...@refinitiv-ui/elements@6.6.1) (2023-01-27)
+
+
+### Bug Fixes
+
+* **list:** error custom renderer requires key ([#575](https://github.com/Refinitiv/refinitiv-ui/issues/575)) ([1a960c9](https://github.com/Refinitiv/refinitiv-ui/commit/1a960c9234fb4a370b60d3901e65b437b1444823))
+* **pagination, tree-select:** remove unused element imports ([#572](https://github.com/Refinitiv/refinitiv-ui/issues/572)) ([0bfc2f5](https://github.com/Refinitiv/refinitiv-ui/commit/0bfc2f5eccd0a6ea50c3927d71d545df20c79d90))
+* **select:** unset cursor of header items ([#571](https://github.com/Refinitiv/refinitiv-ui/issues/571)) ([eff8a75](https://github.com/Refinitiv/refinitiv-ui/commit/eff8a75a68f23348178f05c716db673c97acf2f1))
+
+
+
+
+
+# [6.6.0](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.5.3...@refinitiv-ui/elements@6.6.0) (2023-01-06)
+
+
+### Bug Fixes
+
+* **chart:** change chart property access modifier from private to public ([#565](https://github.com/Refinitiv/refinitiv-ui/issues/565)) ([a25a506](https://github.com/Refinitiv/refinitiv-ui/commit/a25a506e334dead2004ebaca4a815bbf284e9990))
+* **color-dialog:** can not change between default and black color ([#557](https://github.com/Refinitiv/refinitiv-ui/issues/557)) ([27d9f2c](https://github.com/Refinitiv/refinitiv-ui/commit/27d9f2c81336c5cc4ba5bde1b7e5b3918f8c78c8))
+
+
+### Features
+
+* remove alias ([#553](https://github.com/Refinitiv/refinitiv-ui/issues/553)) ([32ce300](https://github.com/Refinitiv/refinitiv-ui/commit/32ce300e6ea0f730abd95f305b15cce406dab1c7))
+
+
+
+
+
 ## [6.5.3](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.5.2...@refinitiv-ui/elements@6.5.3) (2022-12-19)